feat: useGas hook
